requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
dataclasses-json==0.6.7
distro==1.9.0
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.43.0
google-genai==1.55.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.0
langchain-google-genai==4.0.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
markdown-it-py==4.0.0
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlite-vec==0.1.6
tavily-python==0.7.15
tenacity==9.1.2
tiktoken==0.12.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
